Refactored Chrome Tabs initialization logic to port changes to common
